+++ reinforcement_learning/rlclientlib/logger/async_batcher.h
@@ -2,7 +2,7 @@
 
 #include <string>
 
-#include "moving_queue.h"
+#include "event_queue.h"
 #include "ranking_event.h"
 #include "sender.h"
 #include "api_status.h"
