+++ reinforcement_learning/unit_test/event_test.cc
@@ -68,3 +68,21 @@ BOOST_AUTO_TEST_CASE(serialize_empty_ranking)
 
   BOOST_CHECK_EQUAL(serialized.c_str(), expected);
 }
+
+BOOST_AUTO_TEST_CASE(interaction_message_survive_test) {
+  utility::data_buffer buffer;
+  utility::data_buffer expected_buffer;
+  ranking_response resp("interaction_id");
+  resp.push_back(1, 0.1);
+  resp.push_back(2, 0.2);
+  resp.set_chosen_action_id(1);
+
+  ranking_event evt(buffer, "interaction_id", "interaction_context", resp);
+
+  ranking_event expected(expected_buffer, "interaction_id", "interaction_context", resp, 0.75);
+
+  evt.try_drop(0.5, 1);
+  evt.try_drop(0.5, 1);
+
+  BOOST_CHECK_EQUAL(evt.str(), expected.str());
+}
