+++ reinforcement_learning/unit_test/unit_test.vcxproj.filters
@@ -35,9 +35,6 @@
     <ClCompile Include="async_batcher_test.cc">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="concurrent_queue_test.cc">
-      <Filter>Source Files</Filter>
-    </ClCompile>
     <ClCompile Include="live_model_test.cc">
       <Filter>Source Files</Filter>
     </ClCompile>
@@ -92,10 +89,7 @@
     <ClCompile Include="mock_util.cc">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="moving_queue_test.cc">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="message_test.cc">
+    <ClCompile Include="event_queue_test.cc">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
